Move source files into the project root directory
